+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-**UNRELEASED**
+**4.7.0** (2022-04-08)
 
 - Removed all uses of Python 3.5 style ``await yield_()`` from core code and documentation
 - Added tracking of current Asphalt context in a :pep:`555` context variable, available via